Renamed flows directory to legacy_flows and ninja_flows directory to flows
