+++ README.md
@@ -115,7 +115,7 @@ NOTE: You must specify <em>either</em> a yaml file or <em>both</em> a design and
 Alternatively, you can execute the ninja generation step entirely separate from the run step for one or more designs:
 
 <pre>Usage:<code>
-    python bfasst/ninja_flows/ninja_flow_manager.py [--flow FLOW] [--designs DESIGN(S)]
+    python bfasst/flows/ninja_flow_manager.py [--flow FLOW] [--designs DESIGN(S)]
     ninja
 
 options:
