+++ bfasst/flows/vivado_conformal.py
@@ -1,10 +1,10 @@
 """Run vivado, phys_netlist, reverse with xray, then compare with conformal."""
 
-from bfasst.ninja_flows.flow import Flow
+from bfasst.flows.flow import Flow
 from bfasst.ninja_tools.impl.vivado_impl import VivadoImpl
 from bfasst.ninja_tools.compare.conformal.conformal import Conformal
 from bfasst.ninja_tools.rev_bit.xray import Xray
-from bfasst.paths import NINJA_FLOWS_PATH
+from bfasst.paths import FLOWS_PATH
 from bfasst.types import Vendor
 from bfasst.ninja_tools.synth.vivado_synth import VivadoSynth
 
@@ -30,4 +30,4 @@ class VivadoConformal(Flow):
         )
 
     def get_top_level_flow_path(self) -> str:
-        return NINJA_FLOWS_PATH / "vivado_conformal.py"
+        return FLOWS_PATH / "vivado_conformal.py"
