+++ bfasst/flows/vivado_phys_netlist.py
@@ -1,9 +1,9 @@
 """Creates a Vivado netlist that has only physical primitives."""
 
-from bfasst.ninja_flows.flow import Flow
+from bfasst.flows.flow import Flow
 from bfasst.ninja_tools.impl.vivado_impl import VivadoImpl
 from bfasst.ninja_tools.transform.phys_netlist import PhysNetlist
-from bfasst.paths import NINJA_FLOWS_PATH
+from bfasst.paths import FLOWS_PATH
 from bfasst.ninja_tools.synth.vivado_synth import VivadoSynth
 
 
@@ -30,7 +30,7 @@ class VivadoPhysNetlist(Flow):
         # pylint: enable=duplicate-code
 
     def get_top_level_flow_path(self):
-        return NINJA_FLOWS_PATH / "vivado_phys_netlist.py"
+        return FLOWS_PATH / "vivado_phys_netlist.py"
 
     @staticmethod
     def add_required_synth_options(options):
