+++ bfasst/legacy_flows/sub_flows/ic2_impl_ice_rev_one_cmp.py
@@ -1,8 +1,8 @@
 """Subflow for running IC2 Implementation, Icestrom bit reversal, and Onespin compare"""
 
-from bfasst.flows.flow import Flow
-from bfasst.flows.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
-from bfasst.flows.sub_flows.onespin_cmp import OnespinCmp
+from bfasst.legacy_flows.flow import Flow
+from bfasst.legacy_flows.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
+from bfasst.legacy_flows.sub_flows.onespin_cmp import OnespinCmp
 
 
 class Ic2ImplIceRevOneCmp(Flow):
