+++ bfasst/legacy_flows/synplify_ic2_icestorm_onespin.py
@@ -1,36 +1,38 @@
-"""Ic2LseConformal flow"""
+"""SynplifyIc2IcestormOnespin flow"""
 
 # pylint: disable=duplicate-code
 
-from bfasst.flows.flow import Flow
-from bfasst.flows.sub_flows.conformal import Conformal
-from bfasst.flows.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
+from bfasst.compare.onespin import OneSpinCompareTool
+from bfasst.legacy_flows.flow import Flow
+from bfasst.legacy_flows.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
 from bfasst.job import Job
-from bfasst.synth.ic2_lse import Ic2LseSynthesisTool
+from bfasst.synth.ic2_synplify import Ic2SynplifySynthesisTool
 from bfasst.types import ToolType
 
 
-class Ic2LseConformal(Flow):
-    """Ic2LseConformal flow"""
+class SynplifyIc2IcestormOnespin(Flow):
+    """SynplifyIc2IcestormOnespin flow"""
 
     def create(self):
-        """Synthesize and implement using Icecube2 lse, and compare with Conformal"""
+        """
+        Icecube2 Synplify synthesis and implementation, reverse with
+        icestorm, compare with onespin.
+        """
         # Reset job list in case this flow is called multiple times
         self.job_list = []
 
-        lse_synth_tool = Ic2LseSynthesisTool(
+        synplify_synth_tool = Ic2SynplifySynthesisTool(
             self.design.build_dir, self.design, self.flow_args[ToolType.SYNTH]
         )
-        curr_job = Job(lse_synth_tool.create_netlist, self.design.rel_path)
+        curr_job = Job(synplify_synth_tool.create_netlist, self.design.rel_path)
         self.job_list.append(curr_job)
 
-        # Run icecube implementation and icestorm bitstream reversal
         impl_and_rev_sub_flow = Ic2ImplAndIceRev(self.design, self.flow_args)
         impl_and_rev_sub_flow.create()
         impl_and_rev_sub_flow.modify_first_job_dependencies({self.job_list[-1].uuid})
         self.job_list.extend(impl_and_rev_sub_flow.job_list)
 
-        # Set paths for conformal
+        # Set paths for onespin
         self.design.netlist_path = self.design.build_dir / (self.design.top + "_impl.v")
         if self.design.cur_error_flow_name is None:
             self.design.reversed_netlist_path = self.design.build_dir / (
@@ -41,13 +43,19 @@ class Ic2LseConformal(Flow):
                 self.design.top + "_" + self.design.cur_error_flow_name + "_reversed.v"
             )
 
-        # Run conformal
+        # Run onespin
         curr_job = Job(self.adjust_design_object, self.design.rel_path, {self.job_list[-1].uuid})
+        self.job_list.append(curr_job)
 
-        conformal_sub_flow = Conformal(self.design, self.flow_args)
-        conformal_sub_flow.create()
-        conformal_sub_flow.modify_first_job_dependencies({self.job_list[-1].uuid})
-        self.job_list.extend(conformal_sub_flow.job_list)
+        cmp_tool = OneSpinCompareTool(
+            self.design.build_dir,
+            self.design,
+            self.design.netlist_path,
+            self.design.reversed_netlist_path,
+            self.flow_args[ToolType.CMP],
+        )
+        curr_job = Job(cmp_tool.compare_netlists, self.design.rel_path, {self.job_list[-1].uuid})
+        self.job_list.append(curr_job)
 
         return self.job_list
