+++ test/flows/test_vivado_conformal_flow.py
@@ -5,8 +5,8 @@
 
 import unittest
 
-from bfasst.ninja_flows.flow_utils import create_build_file
-from bfasst.ninja_flows.vivado_conformal import VivadoConformal
+from bfasst.flows.flow_utils import create_build_file
+from bfasst.flows.vivado_conformal import VivadoConformal
 from bfasst.ninja_tools.synth.vivado_synth import VivadoSynth
 from bfasst.ninja_tools.impl.vivado_impl import VivadoImpl
 from bfasst.ninja_tools.rev_bit.xray import Xray
@@ -14,7 +14,7 @@ from bfasst.ninja_tools.compare.conformal.conformal import Conformal
 from bfasst.paths import (
     DESIGNS_PATH,
     NINJA_BUILD_PATH,
-    NINJA_FLOWS_PATH,
+    FLOWS_PATH,
 )
 
 
@@ -57,14 +57,14 @@ class TestVivadoConformalFlow(unittest.TestCase):
         VivadoSynth(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
         VivadoImpl(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
         Conformal(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
-        expected.append(NINJA_FLOWS_PATH / "vivado_conformal.py")
+        expected.append(FLOWS_PATH / "vivado_conformal.py")
         observed = sorted([str(s) for s in observed])
         expected = sorted([str(s) for s in expected])
         self.assertEqual(observed, expected)
 
     def test_get_top_level_flow_path(self):
         self.assertEqual(
-            self.flow.get_top_level_flow_path(), NINJA_FLOWS_PATH / "vivado_conformal.py"
+            self.flow.get_top_level_flow_path(), FLOWS_PATH / "vivado_conformal.py"
         )
