+++ test/flows/test_vivado_ooc_flow.py
@@ -2,13 +2,13 @@
 import json
 import unittest
 from bfasst import config
-from bfasst.ninja_flows.vivado_ooc import VivadoOoc
+from bfasst.flows.vivado_ooc import VivadoOoc
 from bfasst.ninja_tools.synth.vivado_synth import VivadoSynth
 from bfasst.ninja_tools.impl.vivado_impl import VivadoImpl
 from bfasst.paths import (
     DESIGNS_PATH,
     NINJA_BUILD_PATH,
-    NINJA_FLOWS_PATH,
+    FLOWS_PATH,
 )
 from bfasst.utils import compare_json
 
@@ -89,8 +89,8 @@ class TestVivadoOocFlow(unittest.TestCase):
         ]
         VivadoSynth(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
         VivadoImpl(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
-        expected.append(NINJA_FLOWS_PATH / "vivado_ooc.py")
-        expected.append(NINJA_FLOWS_PATH / "vivado.py")
+        expected.append(FLOWS_PATH / "vivado_ooc.py")
+        expected.append(FLOWS_PATH / "vivado.py")
 
         observed = sorted([str(s) for s in observed])
         expected = sorted([str(s) for s in expected])
@@ -100,7 +100,7 @@ class TestVivadoOocFlow(unittest.TestCase):
         )
 
     def test_get_top_level_flow_path(self):
-        self.assertEqual(self.flow.get_top_level_flow_path(), NINJA_FLOWS_PATH / "vivado_ooc.py")
+        self.assertEqual(self.flow.get_top_level_flow_path(), FLOWS_PATH / "vivado_ooc.py")
 
 
 if __name__ == "__main__":
