+++ test/flows/test_vivado_phys_netlist.py
@@ -5,15 +5,15 @@
 
 import unittest
 
-from bfasst.ninja_flows.flow_utils import create_build_file
-from bfasst.ninja_flows.vivado_phys_netlist import VivadoPhysNetlist
+from bfasst.flows.flow_utils import create_build_file
+from bfasst.flows.vivado_phys_netlist import VivadoPhysNetlist
 from bfasst.ninja_tools.transform.phys_netlist import PhysNetlist
 from bfasst.ninja_tools.synth.vivado_synth import VivadoSynth
 from bfasst.ninja_tools.impl.vivado_impl import VivadoImpl
 from bfasst.paths import (
     DESIGNS_PATH,
     NINJA_BUILD_PATH,
-    NINJA_FLOWS_PATH,
+    FLOWS_PATH,
 )
 
 
@@ -59,7 +59,7 @@ class TestVivadoPhysNetlist(unittest.TestCase):
         VivadoSynth(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
         VivadoImpl(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
         PhysNetlist(None, DESIGNS_PATH / "byu/alu").add_ninja_deps(expected)
-        expected.append(NINJA_FLOWS_PATH / "vivado_phys_netlist.py")
+        expected.append(FLOWS_PATH / "vivado_phys_netlist.py")
 
         observed = sorted([str(s) for s in observed])
         expected = sorted([str(s) for s in expected])
@@ -67,7 +67,7 @@ class TestVivadoPhysNetlist(unittest.TestCase):
 
     def test_get_top_level_flow_path(self):
         self.assertEqual(
-            self.flow.get_top_level_flow_path(), NINJA_FLOWS_PATH / "vivado_phys_netlist.py"
+            self.flow.get_top_level_flow_path(), FLOWS_PATH / "vivado_phys_netlist.py"
         )
